research: Add research
