+++ research/500-node-from-500-to-1000-app/500-node-from-500-to-1000-app-linear.txt
@@ -1,0 +1,102 @@
+0 5.290989065542817e-06
+10 0.01649237899982836
+20 0.014472563998424448
+30 0.018471289993613027
+40 0.02422608301276341
+50 0.026939054994727485
+60 0.03310399800830055
+70 0.037691973004257306
+80 0.04108171800908167
+90 0.046350858989171684
+100 0.04960792300698813
+110 0.05336915299994871
+120 0.057261369001935236
+130 0.061914878999232315
+140 0.06668536100187339
+150 0.07005592199857347
+160 0.0750301810039673
+170 0.07895166799426079
+180 0.08393452900054399
+190 0.08719909100909717
+200 0.09177732199896127
+210 0.09530807800183538
+220 0.10188156699587125
+230 0.10429815200041048
+240 0.1082715139928041
+250 0.12321624600735959
+260 0.11677569399762433
+270 0.12127534199680667
+280 0.125668794993544
+290 0.12910188200476114
+300 0.13389772099617403
+310 0.1393675920116948
+320 0.14213093199941795
+330 0.14556504599750042
+340 0.15092289198946673
+350 0.15525566699216142
+360 0.15986605800571851
+370 0.16315557999769226
+380 0.16745193899259903
+390 0.17189839099592064
+400 0.17668494400277268
+410 0.1799216420040466
+420 0.1951674070005538
+430 0.1901381420029793
+440 0.19377424800768495
+450 0.19804874699912034
+460 0.2015507769974647
+470 0.20578260799811687
+480 0.21147101699898485
+490 0.2139155810000375
+500 0.21878673500032164
+510 0.2234395220002625
+520 0.22837554800207727
+530 0.23108187499747146
+540 0.23603755100339185
+550 0.24103881799965166
+560 0.24288600499858148
+570 0.2483076600037748
+580 0.25719463599671144
+590 0.25559582900314126
+600 0.2607810409972444
+610 0.2664919599919813
+620 0.27191613300237805
+630 0.27268525600084104
+640 0.2847408249945147
+650 0.2827665530057857
+660 0.28559949299960863
+670 0.28904549599974416
+680 0.29433152099954896
+690 0.2964602079882752
+700 0.30161539500113577
+710 0.3124219260062091
+720 0.31473241599451285
+730 0.31594211800256744
+740 0.3203994389914442
+750 0.32687826101027895
+760 0.3290181899938034
+770 0.3341429490101291
+780 0.33546975400531664
+790 0.34262012400722597
+800 0.3433705430070404
+810 0.3469389020028757
+820 0.3530872260016622
+830 0.3586553269997239
+840 0.3639984259934863
+850 0.3712067830056185
+860 0.3726243769924622
+870 0.3760653089993866
+880 0.380052730004536
+890 0.38336343300761655
+900 0.38758823300304357
+910 0.39149914198787883
+920 0.39456708800571505
+930 0.3971275119984057
+940 0.406868129997747
+950 0.40718535700580105
+960 0.41185535700060427
+970 0.41864981199614704
+980 0.4219766280002659
+990 0.42749377799918875
+1000 0.4304378599917982
+
