+++ research/500-node-from-500-to-1000-app/500-node-from-500-to-1000-app.txt
@@ -1,50 +1,63 @@
-500	2.0080794559	22.852071285
-510	1.9901637791	24.0996925849
-520	2.015780576	25.0351794129
-530	2.2137415202	31.4076489091
-540	2.1047025679	29.5419783511
-550	2.17792883	30.418083804
-560	2.2938713601	30.4675314089
-570	2.2064662292	31.1789705809
-580	2.2023421819	31.932663074
-590	2.2229970461	32.31220072
-600	2.2546242801	32.1040917742
-610	2.269058699	34.1318329689
-620	2.3129963451	32.9856162211
-630	2.3720989949	29.213524011
-640	3.1225938001	25.892144803
-650	3.46677792	24.7125960879
-660	2.6508954661	27.817391474
-670	2.7166316819	34.822335318
-680	2.5371136311	28.0386159888
-690	3.0193468281	28.902043019
-700	2.5736568491	28.3945765861
-710	2.7159298309	27.7418197962
-720	3.4957207481	29.096944496
-730	3.328773815	31.605051863
-740	3.3055704818	34.5520706892
-750	3.5872661669	32.3495212079
-760	3.3044123959	31.1973445602
-770	3.1510633251	33.566291438
-780	3.0106374701	31.9064460921
-790	3.4750703669	30.6640042579
-800	3.0404921891	39.9170019391
-810	3.271606684	41.038757382
-820	3.857777514	34.298836811
-830	4.2217126561	30.525901098
-840	4.4745547511	32.2068285029
-850	4.726091411	34.248486418
-860	4.3126773119	31.2976478371
-870	3.363333239	31.97575479
-880	3.8352678679	32.6425986001
-890	3.403653298	32.4150151359
-900	3.7778523469	33.7108376431
-910	3.4995879009	34.0570139631
-920	3.3933716889	34.4780712109
-930	3.3706396301	34.344213614
-940	3.433433613	34.8930626311
-950	3.4452877461	35.3591244761
-960	3.5470172989	34.910235737
-970	3.7427881099	37.620928965
-980	3.7067529368	39.525735925
-990	3.794830617	48.8376154671
+0 0.21983525999530684
+10 0.6799951399880229
+20 1.0533640119974734
+30 1.4563150699977996
+40 1.826419558012276
+50 2.2375032609998016
+60 2.6243447070010006
+70 3.009104039010708
+80 3.3563503659970593
+90 3.7699531460093567
+100 4.1209243230114225
+110 4.753774683005759
+120 5.420805007001036
+130 5.830313981001382
+140 6.086042454990093
+150 6.63942986400798
+160 7.712794467995991
+170 8.175109203002648
+180 8.705328919997555
+190 9.575077243993292
+200 10.448863120996975
+210 11.506570631987415
+220 12.761861745006172
+230 13.647163849993376
+240 14.937281061997055
+250 15.991384915003437
+260 17.37818195699947
+270 18.76855218000128
+280 20.524561926999013
+290 21.919441634003306
+300 23.27901039599965
+310 25.116848040008335
+320 26.88378917099908
+330 28.68450958300673
+340 31.041425875009736
+350 34.041812825002125
+360 35.853292890999
+370 37.63565087701136
+380 39.69058332599525
+390 42.43002656499448
+400 43.8967175459984
+410 45.923517664996325
+420 48.7483977029915
+430 51.97317415100406
+440 53.5748801860027
+450 57.30050984599802
+460 59.738434844999574
+470 61.216721835007775
+480 64.65944619500078
+490 70.40088924199517
+500 71.71061150199967
+510 73.33580020700174
+520 76.43537309599924
+530 84.29040821599483
+540 83.1649913170113
+550 87.1707402930042
+560 91.46100770200428
+570 93.8735884130001
+580 96.6497876009962
+590 99.59425126599672
+600 101.97642188899044
+610 107.65081501301029
+620 111.22452751899255
