+++ research/centralized/centralized_response_times.data
@@ -1,0 +1,100 @@
+0.0	100.5
+0.01	100.5
+0.02	100.5
+0.03	100.5
+0.04	100.5
+0.05	100.5
+0.06	100.5
+0.07	100.5
+0.08	100.5
+0.09	100.5
+0.1	100.5
+0.11	100.5
+0.12	100.5
+0.13	100.5
+0.14	100.5
+0.15	100.5
+0.16	100.5
+0.17	100.5
+0.18	100.5
+0.19	100.5
+0.2	100.5
+0.21	100.5
+0.22	100.5
+0.23	100.5
+0.24	100.5
+0.25	100.5
+0.26	100.5
+0.27	100.5
+0.28	100.5
+0.29	100.5
+0.3	100.5
+0.31	100.5
+0.32	100.5
+0.33	100.5
+0.34	100.5
+0.35	100.5
+0.36	100.5
+0.37	100.5
+0.38	100.5
+0.39	100.5
+0.4	100.5
+0.41	100.5
+0.42	100.5
+0.43	100.5
+0.44	100.5
+0.45	100.5
+0.46	100.5
+0.47	100.5
+0.48	100.5
+0.49	100.5
+0.5	100.5
+0.51	100.5
+0.52	100.5
+0.53	100.5
+0.54	100.5
+0.55	100.5
+0.56	100.5
+0.57	100.5
+0.58	100.5
+0.59	100.5
+0.6	100.5
+0.61	100.5
+0.62	100.5
+0.63	100.5
+0.64	100.5
+0.65	100.5
+0.66	100.569893156
+0.67	100.747008372
+0.68	100.925034691
+0.69	101.106480482
+0.7	101.29367514
+0.71	101.490530981
+0.72	101.683839677
+0.73	101.866677146
+0.74	102.061300153
+0.75	102.248640767
+0.76	102.425730977
+0.77	102.606044636
+0.78	102.791082271
+0.79	102.982628958
+0.8	103.174852764
+0.81	103.387718912
+0.82	103.593744107
+0.83	103.813106852
+0.84	104.032607364
+0.85	104.248138527
+0.86	104.479666333
+0.87	104.738765831
+0.88	104.98471937
+0.89	105.2485032
+0.9	105.527046523
+0.91	105.83705599
+0.92	106.183396979
+0.93	106.578329737
+0.94	107.051453259
+0.95	107.6353006
+0.96	108.286237616
+0.97	109.14204046
+0.98	110.126394842
+0.99	111.324522185
